\name{plot.ts.default}
\alias{plot.ts}
\title{Default function to plot a time series.
}
\usage{
plot.ts(data, var, ...)
}
\arguments{
  \item{data}{A data table e.g. from read.loggerdat}
  \item{var}{The columnname of the column that should be plotted}
  \item{\dots}{supplementary arguments (to Do)
}
}
\value{A plot is generated
}
\author{
}
\note{To Do: writing xlab and ylab , scaling of xaxis, getting a windrose when plotting wind
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
